backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   starlette
    #   watchfiles
click==8.1.7
    # via uvicorn
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.4
    # via smart_llm_tools (pyproject.toml)
h11==0.14.0
    # via uvicorn
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via anyio
itsdangerous==2.2.0
    # via smart_llm_tools (pyproject.toml)
pydantic==2.9.2
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.6.1
    # via smart_llm_tools (pyproject.toml)
pyjwt==2.9.0
    # via smart_llm_tools (pyproject.toml)
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.2
    # via uvicorn
sniffio==1.3.1
    # via anyio
starlette==0.41.2
    # via fastapi
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   uvicorn
uvicorn[standard]==0.32.0
    # via smart_llm_tools (pyproject.toml)
uvloop==0.21.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
